상세검색 임시저장

+++ src/main/java/controller/myPage/deleteController.java
@@ -1,7 +1,9 @@
 package controller.myPage;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import javax.servlet.ServletException;
@@ -11,29 +13,34 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 
+import data.Like;
+import data.User;
+import repository.PostDAO;
 import repository.UserDAO;
 
 @WebServlet("/user/delete")
-public class deleteController extends HttpServlet{
-	
+public class deleteController extends HttpServlet {
+
 	@Override
 	protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-		
-		
+
 		HttpSession session = req.getSession();
-//		Boolean logon = (Boolean)session.getAttribute("logon");
-//		if(logon == null || !logon) {
-//			resp.sendRedirect("/user/login");
-//			return;
-//		}
-		
+		User logonUser = (User) session.getAttribute("logonUser");
+		Boolean logon = (Boolean) session.getAttribute("logon");
+		if (logon == null || !logon) {
+			resp.sendRedirect("/user/login");
+			return;
+		}
+
 		String id = req.getParameter("id");
 		
+
+		PostDAO.allDeleteCamp(id);
+		PostDAO.deleteReview(id);
 		UserDAO.delete(id);
-		
+
 		session.invalidate();
-		
-		
+
 		req.getRequestDispatcher("/WEB-INF/views/index.jsp").forward(req, resp);
 	}
